+++ src/api/.env
@@ -1,6 +1,6 @@
 DB_CONNECTION_STRING=sqlite:///../../sqlite/raj.db
 FLASK_APP_KEY="any key works"
 JWT_SECRET_KEY="a good random one"
-FLASK_APP=src/main.py
+FLASK_APP=main.py
 FLASK_ENV=development
 FLASK_PORT=3000
