+++ src/api/backend/access_control.py
@@ -1,7 +1,7 @@
 from os import environ
 from flask_jwt_extended import JWTManager
-from api import app
-from api.models import User
+from backend import app
+from backend.models import User
 
 app.config["JWT_SECRET_KEY"] = environ.get("JWT_SECRET_KEY")
 jwt = JWTManager(app)
