+++ src/api/backend/endpoints/endpoints_authentication.py
@@ -1,0 +1,4 @@
+from backend import app
+from backend.controllers.authentication_controllers import *
+
+app.add_url_rule("/login/", methods=["POST"], view_func=login)
