+++ src/api/backend/endpoints/endpoints_user.py
@@ -1,5 +1,5 @@
-from api import app
-from api.controllers.users_controllers import *
+from backend import app
+from backend.controllers.users_controllers import *
 
 app.add_url_rule("/users/", methods=["GET"], view_func=get_all_users)
 app.add_url_rule("/users/<int:user_id>/", methods=["GET"], view_func=get_user_by_id)
