--- src/api/src/api/endpoints/endpoints_authentication.py
@@ -1,4 +1,0 @@
-from api import app
-from api.controllers.authentication_controllers import *
-
-app.add_url_rule("/login/", methods=["POST"], view_func=login)
